Move snapshot tests to test/snapshot
