+++ test/snapshot/lib.js
@@ -1,9 +1,9 @@
 const path = require('path');
 const fs = require('fs');
-const VM = require('../src/virtual-machine');
-const JSGenerator = require('../src/compiler/jsgen');
+const VM = require('../../src/virtual-machine');
+const JSGenerator = require('../../src/compiler/jsgen');
 
-const executeDir = path.resolve(__dirname, '../test/fixtures/execute');
+const executeDir = path.resolve(__dirname, '../fixtures/execute');
 // sb2 project loading results in random IDs each time, so for now we only snapshot sb3 files
 const testProjects = fs.readdirSync(executeDir).filter(uri => uri.endsWith('.sb3'));
